add:start speech

+++ speaker.h
@@ -1,9 +1,11 @@
 #pragma once
+
 #include "iostream"
+
 using namespace std;
 
 class Speaker {
 public:
     string m_Name;
-    double m_Score[2];
+    float m_Score[2];
 };
